requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
build==1.0.3
    # via pip-tools
click==8.1.7
    # via pip-tools
jinja2==3.1.3
    # via -r requirements.in
markupsafe==2.1.4
    # via jinja2
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pyyaml==6.0.1
    # via -r requirements.in
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
